LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/browsers/the-window-object/open-close/open_initial_size-expected.txt


FAIL Window geometry is available in opener context immediately assert_equals: expected 980 but got 800
